feat(docs-site): seo optimization
