+++ apps/docs-site/src/app/docs/libraries/versioning/page.tsx
@@ -1,4 +1,10 @@
+import type { Metadata } from 'next'
 import { LibraryDocPage } from '@/components/library-doc-page'
+import { getLibraryMetadata } from '@/lib/metadata'
+
+export function generateMetadata(): Metadata {
+  return getLibraryMetadata('versioning')
+}
 
 export default function VersioningPage() {
   return (
