+++ apps/docs-site/src/app/docs/page.tsx
@@ -1,8 +1,14 @@
+import type { Metadata } from 'next'
 import { Breadcrumb } from '@/components/breadcrumb'
 import { CodeBlock } from '@/components/code-block'
 import { DocsContentWrapper } from '@/components/docs-content-wrapper'
 import { H1, H2 } from '@/components/heading-with-anchor'
 import Link from 'next/link'
+
+export const metadata: Metadata = {
+  title: 'Getting Started',
+  description: 'Set up hyperfrontend and create your first micro-frontend feature in minutes with Nx workspace integration.',
+}
 
 export default function DocsPage() {
   return (
