+++ apps/docs-site/src/app/layout.tsx
@@ -2,9 +2,14 @@ import type { Metadata } from 'next'
 import { ThemeProvider } from '@/components/theme-provider'
 import '@/lib/dev-logger'
 import '@/styles/globals.css'
+import { createURL } from '@hyperfrontend/immutable-api-utils/built-in-copy/url'
 
 export const metadata: Metadata = {
-  title: 'HyperFrontend - Micro-Frontend Architecture',
+  metadataBase: createURL('https://hyperfrontend.dev'),
+  title: {
+    default: 'HyperFrontend - Micro-Frontend Architecture',
+    template: '%s | HyperFrontend',
+  },
   description:
     'A hybrid micro-frontend pattern to embed live web applications with communication protocols, lifecycle, and contract standards.',
   keywords: ['micro-frontend', 'microfrontend', 'web components', 'iframe', 'react', 'angular', 'vue', 'svelte'],
@@ -15,6 +20,16 @@ export const metadata: Metadata = {
     url: 'https://hyperfrontend.dev',
     siteName: 'HyperFrontend',
     type: 'website',
+    locale: 'en_US',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'HyperFrontend',
+    description: 'Compose your existing apps together securely — like Lego bricks.',
+    creator: '@andrew_redican',
+  },
+  alternates: {
+    canonical: './',
   },
 }
